+++ core/src/main/java/org/mskcc/cbio/cgds/model/CnaEvent.java
@@ -21,7 +21,7 @@ public class CnaEvent {
     public static enum CNA {
         AMP ((short)2, "Amplified"),
         GAIN ((short)1, "Gained"),
-        HETLOSS ((short)-1, "Hemizygously deleted"),
+        HETLOSS ((short)-1, "Heterozygously deleted"),
         HOMDEL ((short)-2, "Homozygously deleted");
         
         private short code;
@@ -95,12 +95,7 @@ public class CnaEvent {
     }
     
     public String getGeneSymbol() {
-        try {
-            return DaoGeneOptimized.getInstance().getGene(entrezGeneId).getHugoGeneSymbolAllCaps();
-        } catch (DaoException ex) {
-            ex.printStackTrace();
-            return null;
-        }
+        return DaoGeneOptimized.getInstance().getGene(entrezGeneId).getHugoGeneSymbolAllCaps();
     }
 
     public void setEntrezGeneId(long entrezGeneId) {