+++ core/src/main/java/org/mskcc/cbio/cgds/scripts/ResetDatabase.java
@@ -70,8 +70,7 @@ public class ResetDatabase {
         daoClinicalAttribute.deleteAllRecords();
         DaoClinicalFreeForm daoClinicalFreeForm = new DaoClinicalFreeForm();
         daoClinicalFreeForm.deleteAllRecords();
-        DaoMutation daoMutation = DaoMutation.getInstance();
-        daoMutation.deleteAllRecords();
+        DaoMutation.deleteAllRecords();
         DaoMutationFrequency daoMutationFrequency = new DaoMutationFrequency();
         daoMutationFrequency.deleteAllRecords();
         DaoGeneticProfileCases daoGeneticProfileCases = new DaoGeneticProfileCases();