+++ core/src/main/java/org/mskcc/cbio/portal/servlet/CnaJSON.java
@@ -14,7 +14,6 @@ import org.json.simple.JSONValue;
 import org.mskcc.cbio.cgds.dao.*;
 import org.mskcc.cbio.cgds.model.CancerStudy;
 import org.mskcc.cbio.cgds.model.CanonicalGene;
-import org.mskcc.cbio.cgds.model.Case;
 import org.mskcc.cbio.cgds.model.CnaEvent;
 import org.mskcc.cbio.cgds.model.CopyNumberSegment;
 import org.mskcc.cbio.cgds.model.GeneticProfile;
@@ -63,6 +62,7 @@ public class CnaJSON extends HttpServlet {
             throws ServletException, IOException {
         String patient = request.getParameter(PatientView.PATIENT_ID);
         String cnaProfileId = request.getParameter(PatientView.CNA_PROFILE);
+        String mrnaProfileId = request.getParameter(PatientView.MRNA_PROFILE);
         String drugType = request.getParameter(PatientView.DRUG_TYPE);
         boolean fdaOnly = false;
         boolean cancerDrug = true;
@@ -74,32 +74,38 @@ public class CnaJSON extends HttpServlet {
         GeneticProfile cnaProfile;
         CancerStudy cancerStudy = null;
         List<CnaEvent> cnaEvents = Collections.emptyList();
-        Map<String, Set<String>> drugs = Collections.emptyMap();
+        Map<Long, Set<String>> drugs = Collections.emptyMap();
         Map<Long, Integer>  contextMap = Collections.emptyMap();
+        Map<Long, Map<String,Object>> mrnaContext = Collections.emptyMap();
 
         try {
             cnaProfile = DaoGeneticProfile.getGeneticProfileByStableId(cnaProfileId);
             cancerStudy = DaoCancerStudy.getCancerStudyByInternalId(cnaProfile.getCancerStudyId());
-            if (cnaProfile!=null) {
-                cnaEvents = DaoCnaEvent.getCnaEvents(patient, cnaProfile.getGeneticProfileId());
-                String concatEventIds = getConcatEventIds(cnaEvents);
-                int profileId = cnaProfile.getGeneticProfileId();
-                drugs = getDrugs(cnaEvents, fdaOnly, cancerDrug);
-                contextMap = DaoCnaEvent.countSamplesWithCnaEvents(concatEventIds, profileId);
+            cnaEvents = DaoCnaEvent.getCnaEvents(patient, cnaProfile.getGeneticProfileId());
+            String concatEventIds = getConcatEventIds(cnaEvents);
+            int profileId = cnaProfile.getGeneticProfileId();
+            drugs = getDrugs(cnaEvents, fdaOnly, cancerDrug);
+            contextMap = DaoCnaEvent.countSamplesWithCnaEvents(concatEventIds, profileId);
+            if (mrnaProfileId!=null) {
+                mrnaContext = getMrnaContext(patient, cnaEvents, mrnaProfileId);
             }
         } catch (DaoException ex) {
             throw new ServletException(ex);
         }
         
         Map<String,List> data = initMap();
+        DaoGeneOptimized daoGeneOptimized = DaoGeneOptimized.getInstance();
         for (CnaEvent cnaEvent : cnaEvents) {
             Set<String> drug = Collections.emptySet();
             try {
-                drug = drugs.get(cnaEvent.getGeneSymbol());
+                drug = drugs.get(cnaEvent.getEntrezGeneId());
             } catch (Exception e) {
                 logger.error(e.getMessage());
             }
-            exportCnaEvent(data, cnaEvent, cancerStudy, drug, contextMap.get(cnaEvent.getEventId()));
+            exportCnaEvent(data, cnaEvent, cancerStudy, drug,
+                    contextMap.get(cnaEvent.getEventId()),
+                    mrnaContext.get(cnaEvent.getEntrezGeneId()),
+                    daoGeneOptimized);
         }
 
         response.setContentType("application/json");
@@ -182,7 +188,7 @@ public class CnaJSON extends HttpServlet {
         return sb.toString();
     }
     
-    private Map<String, Set<String>> getDrugs(List<CnaEvent> cnaEvents, boolean fdaOnly, boolean cancerDrug)
+    private Map<Long, Set<String>> getDrugs(List<CnaEvent> cnaEvents, boolean fdaOnly, boolean cancerDrug)
             throws DaoException {
         DaoDrugInteraction daoDrugInteraction = DaoDrugInteraction.getInstance();
         Set<Long> genes = new HashSet<Long>();
@@ -213,11 +219,9 @@ public class CnaJSON extends HttpServlet {
         }
         
         Map<Long, List<String>> map = daoDrugInteraction.getDrugs(genes,fdaOnly,cancerDrug);
-        Map<String, Set<String>> ret = new HashMap<String, Set<String>>(map.size());
+        Map<Long, Set<String>> ret = new HashMap<Long, Set<String>>(map.size());
         for (Map.Entry<Long, List<String>> entry : map.entrySet()) {
-            String symbol = DaoGeneOptimized.getInstance().getGene(entry.getKey())
-                    .getHugoGeneSymbolAllCaps();
-            ret.put(symbol, new HashSet<String>(entry.getValue()));
+            ret.put(entry.getKey(), new HashSet<String>(entry.getValue()));
         }
         
         // Temporary way of handling cases such as akt inhibitor for pten loss
@@ -225,12 +229,10 @@ public class CnaJSON extends HttpServlet {
             Set<Long> eventGenes = mapTargetToEventGenes.get(entry.getKey());
             if (eventGenes!=null) {
                 for (long eventGene : eventGenes) {
-                    String symbol = DaoGeneOptimized.getInstance().getGene(eventGene)
-                        .getHugoGeneSymbolAllCaps();
-                    Set<String> drugs = ret.get(symbol);
+                    Set<String> drugs = ret.get(eventGene);
                     if (drugs==null) {
                         drugs = new HashSet<String>();
-                        ret.put(symbol, drugs);
+                        ret.put(eventGene, drugs);
                     }
                     drugs.addAll(entry.getValue());
                 }
@@ -241,29 +243,79 @@ public class CnaJSON extends HttpServlet {
         return ret;
     }
     
+    private Map<Long, Map<String,Object>> getMrnaContext(String caseId, List<CnaEvent> cnaEvents,
+            String mrnaProfileId) throws DaoException {
+        Map<Long, Map<String,Object>> mapGenePercentile = new HashMap<Long, Map<String,Object>>();
+        DaoGeneticAlteration daoGeneticAlteration = DaoGeneticAlteration.getInstance();
+        for (CnaEvent cnaEvent : cnaEvents) {
+            long gene = cnaEvent.getEntrezGeneId();
+            if (mapGenePercentile.containsKey(gene)) {
+                continue;
+            }
+            
+            Map<String,String> mrnaMap = daoGeneticAlteration.getGeneticAlterationMap(
+                    DaoGeneticProfile.getGeneticProfileByStableId(mrnaProfileId).getGeneticProfileId(),
+                    gene);
+            double mrnaCase = parseNumber(mrnaMap.get(caseId));
+            if (Double.isNaN(mrnaCase)) {
+                continue;
+            }
+            
+            Map<String,Object> map = new HashMap<String,Object>();
+            mapGenePercentile.put(gene, map);
+            
+            map.put("zscore", mrnaCase);
+            
+            int total = 0, below = 0;
+            for (String strMrna : mrnaMap.values()) {
+                double mrna = parseNumber(strMrna);
+                if (Double.isNaN(mrna)) {
+                    continue;
+                }
+                
+                total++;
+                if (mrna <= mrnaCase) {
+                    below++;
+                }
+            }
+            
+            map.put("perc", 100*below/total);
+        }
+        
+        return mapGenePercentile;
+    }
+    
+    private double parseNumber(String mrna) {
+        try {
+            return Double.parseDouble(mrna);
+        } catch (Exception e) {
+            return Double.NaN;
+        }
+    }
+    
     private Map<String,List> initMap() {
         Map<String,List> map = new HashMap<String,List>();
         map.put("id", new ArrayList());
         map.put("entrez", new ArrayList());
         map.put("gene", new ArrayList());
         map.put("alter", new ArrayList());
+        map.put("mrna", new ArrayList());
         map.put("gistic", new ArrayList());
         map.put("sanger", new ArrayList());
-        map.put("impact", new ArrayList());
+        map.put("cancer-gene", new ArrayList());
         map.put("drug", new ArrayList());
         map.put("altrate", new ArrayList());
         return map;
     }
     
     private void exportCnaEvent(Map<String,List> data, CnaEvent cnaEvent,
-            CancerStudy cancerStudy, Set<String> drugs, Integer context) 
+            CancerStudy cancerStudy, Set<String> drugs, Integer context,
+            Map<String,Object> mrna, DaoGeneOptimized daoGeneOptimized) 
             throws ServletException {
         String symbol = null;
         try {
-            symbol = DaoGeneOptimized.getInstance().getGene(cnaEvent.getEntrezGeneId())
+            symbol = daoGeneOptimized.getGene(cnaEvent.getEntrezGeneId())
                     .getHugoGeneSymbolAllCaps();
-        } catch (DaoException ex) {
-            throw new ServletException(ex);
         } catch (Exception ex) {
             logger.error(ex.getMessage());
             return;
@@ -272,6 +324,7 @@ public class CnaJSON extends HttpServlet {
         data.get("gene").add(symbol);
         data.get("entrez").add(cnaEvent.getEntrezGeneId());
         data.get("alter").add(cnaEvent.getAlteration().getCode());
+        data.get("mrna").add(mrna);
         
         // TODO: GISTIC
         List gistic;
@@ -286,15 +339,15 @@ public class CnaJSON extends HttpServlet {
         data.get("altrate").add(context);
         
         boolean isSangerGene = false;
-        boolean isIMPACTGene = false;
+        boolean isCbioCancerGene = false;
         try {
             isSangerGene = DaoSangerCensus.getInstance().getCancerGeneSet().containsKey(symbol);
-            isIMPACTGene = DaoGeneOptimized.getInstance().isIMPACTGene(symbol);
+            isCbioCancerGene = daoGeneOptimized.isCbioCancerGene(symbol);
         } catch (DaoException ex) {
             throw new ServletException(ex);
         }
         data.get("sanger").add(isSangerGene);
-        data.get("impact").add(isIMPACTGene);
+        data.get("cancer-gene").add(isCbioCancerGene);
         
         // drug
         data.get("drug").add(drugs);