+++ core/src/main/java/org/mskcc/cbio/portal/util/SkinUtil.java
@@ -35,7 +35,7 @@ import org.mskcc.cbio.portal.servlet.QueryBuilder;
  * @author Ethan Cerami.
  */
 public class SkinUtil {
-    public static final String DEFAULT_TITLE = "cBio Cancer Genomics Portal";
+    public static final String DEFAULT_TITLE = "cBioPortal for Cancer Genomics";
     public static final String DEFAULT_EMAIL_CONTACT = "cbioportal at googlegroups dot com";
     public static final String DEFAULT_AUTHORIZATION_MESSAGE = "Access to this portal is only " +
             "available to authorized users.";
@@ -245,7 +245,7 @@ public class SkinUtil {
     }
     
     public static String getLinkToPatientView(String caseId, String cancerStudyId) {
-        return "tumormap.do?" + org.mskcc.cbio.portal.servlet.PatientView.PATIENT_ID
+        return "case.do?" + org.mskcc.cbio.portal.servlet.PatientView.PATIENT_ID
                 + "=" + caseId + "&" + QueryBuilder.CANCER_STUDY_ID + "=" + cancerStudyId;
     }