+++ mondrian-cy3-app/src/main/java/org/mskcc/mondrian/internal/MondrianUtil.java
@@ -1,9 +1,0 @@
-package org.mskcc.mondrian.internal;
-
-import org.cytoscape.model.CyNetwork;
-
-public class MondrianUtil {
-	public static CyNetwork getCurrentNetwork() {
-		return MondrianApp.getInstance().getAppManager().getCurrentNetwork();
-	}
-}