+++ portal/src/main/webapp/js/oncoprint.js
@@ -146,7 +146,7 @@ var _Oncoprint = function(wrapper, params) {
     var TRACK_PADDING = 4;
     var LABEL_PADDING = 40;
     var LITTLE_RECT_HEIGHT = RECT_HEIGHT / 3;
-    var MRNA_STROKE_WIDTH = 1;
+    var MRNA_STROKE_WIDTH = 2;
     var UPREGULATED = "UPREGULATED";
     var DOWNREGULATED = "DOWNREGULATED";
     var MRNA_UP_COLOR = "#FF9999";
@@ -377,7 +377,7 @@ var _Oncoprint = function(wrapper, params) {
                 AMPLIFIED: "Amplification",
                 GAINED: "Gain",
                 DIPLOID: "Diploid",
-                HEMIZYGOUSLYDELETED: "Hemizygous Deletion",
+                HEMIZYGOUSLYDELETED: "Heterozygous Deletion",
                 HOMODELETED: "Homozygous Deletion"
             },
             mrna: {
@@ -535,7 +535,7 @@ var _Oncoprint = function(wrapper, params) {
 
         var patientViewUrl = function(sample_id) {
             // helper function
-            var href = "tumormap.do?case_id=" + sample_id
+            var href = "case.do?case_id=" + sample_id
                 + "&cancer_study_id=" + params.cancer_study_id;
 
             return "<a href='" + href + "'>" + sample_id + "</a>";
@@ -794,7 +794,11 @@ var _Oncoprint = function(wrapper, params) {
             .append(tracks);
 //            .append(visualizedKeys(query.data_types));
 
-        return (new XMLSerializer()).serializeToString(export_svg[0]);
+        return (new XMLSerializer()).serializeToString(export_svg[0])
+            .replace(' xmlns="http://www.w3.org/1999/xhtml"', '');
+        // Firefox and safari implementations of XMLSerializer are different.
+        // For some reason (actually a very good one) they think that this is
+        // XHTML and give it the proper namespace.
     };
 
     return that;