+++ portal/src/main/resources/content/news.markdown
@@ -1,3 +1,17 @@
+# April 28, 2013
+
+* All TCGA data updated to the March Firehose run (March 26, 2012).
+* mRNA percentiles for altered genes shown in patient view.
+
+# April 2, 2013
+
+* All TCGA data updated to the February Firehose run (February 22, 2012).
+
+# March 28, 2013
+
+* All TCGA data updated to the January Firehose run (January 16, 2012).
+* Data from a new bladder cancer study from MSKCC has been added (97 samples, Iyer et al., JCO in press).
+
 # February 16, 2013
 
 * The cBio Portal now contains mutation data from all provisional TCGA projects. Please adhere to [the TCGA publication guidelines](http://cancergenome.nih.gov/abouttcga/policies/publicationguidelines) when using these and any TCGA data in your publications.